+++ koala-bear/src/lib.rs
@@ -5,10 +5,14 @@ extern crate alloc;
 #[cfg(test)]
 mod extension_test;
 mod koala_bear;
+mod packed_quintic_extension;
 mod poseidon2;
+mod quintic_extension;
 
 pub use koala_bear::*;
 pub use poseidon2::*;
+pub use quintic_extension::*;
+pub use packed_quintic_extension::*;
 
 #[cfg(all(target_arch = "aarch64", target_feature = "neon"))]
 mod aarch64_neon;
